Nightly updates (runs but doesn't work properly)
Still has a problem where the mips of the envTexture are not loaded properly. The background only uses mip 0, but the spheres start using more mips (and are incorrect). If you change the sampler in WriteEnvDescriptors() to only use mip 1, you can see the errors in the background as well.