Experiment 101 Metal running
The code now exists to render the rotating cube but it's not working properly. It seems like the vertex buffer is improperly offset (assuming float4s instead of float3s) but when I try to create a vertexDescriptor for the pipeline state I can't even view it in the debugger.