Start using NS::SharedPtr to follow metal-cpp's memory allocations
Change all the allocated ObjC pointers to be stored as NS::SharedPtr() and take ownership from the metal-cpp library using NS::TransferPtr().

Make sure to check every metal-cpp allocation for nullptr return values (which  we would have missed in the past because calling ObjC functions on nullptrs is considered valid behavior).

Remove the allocations for VertexAttribute and VertexLayout descriptors as they are already allocated as part of the VertexDescriptor initialization.