Fix descriptor set counts
Switch from 2 to 1 descriptor set count, which was causing runtime debug layer errors. With that fixed the
program now runs and uses descriptor sets instead of a descriptor buffer.

I also added a Feature flag to enable Descriptor Buffers (which defaults to true) and is set to false for
this graphics experiment.