Initial Agility SDK Checkin
Setup the required exported variables for using the Agility SDK based on the new -DGREX_AGILITY_SDK=X.Y.Z command line
parameter. The way to use it is in README.md but does not actually call NuGet to get the package in question, instead
giving instructions on what steps you'll need to take in Visual Studio.